refactory test

+++ test/models/user.go
@@ -4,6 +4,6 @@ type User struct {
 	ID        string `bson:"_id"`
 	Name      string `bson:"name"`
 	Age       uint16 `bson:"age"`
-	CreatedAt uint16 `bson:"created_at"`
-	UpdatedAt uint16 `bson:"updated_at"`
+	CreatedAt uint64 `bson:"created_at"`
+	UpdatedAt uint64 `bson:"updated_at"`
 }
